requirements.txt
nvidia_ml_py==12.570.86
pyusb==1.3.1
